+++ R/MsBackendMzR.R
@@ -83,12 +83,14 @@ setMethod("peaksData", "MsBackendMzR", function(object) {
     if (!length(object))
         return(list())
     fls <- unique(object@spectraData$dataStorage)
+    read_header <- getOption("READ_HEADER", FALSE)
     if (length(fls) > 1) {
         f <- factor(dataStorage(object), levels = fls)
         unsplit(mapply(FUN = .mzR_peaks, fls, split(scanIndex(object), f),
+                       MoreArgs = list(readHeader = read_header),
                        SIMPLIFY = FALSE, USE.NAMES = FALSE), f)
     } else
-        .mzR_peaks(fls, scanIndex(object))
+        .mzR_peaks(fls, scanIndex(object), readHeader = read_header)
 })
 
 #' @rdname hidden_aliases